feat: handling error in delete user method

+++ users-api/src/main/java/com/github/nmsilos/usersapi/service/UserService.java
@@ -37,6 +37,9 @@ public class UserService {
 
     @Transactional
     public void delete(Long id) {
+        if (!userRepository.existsById(id)) {
+            throw new EntityNotFoundException(String.format("User with id '%s' not found", id));
+        }
         userRepository.deleteById(id);
     }
